extract_trailer_index returns uint32_t

+++ include/extract_trailer_index.h
@@ -1,4 +1,4 @@
-constexpr const size_t extract_trailer_index(const char* const cstr, const bool reversing = true, size_t count = 0) {
+constexpr const uint32_t extract_trailer_index(const char* const cstr, const bool reversing = true, size_t count = 0) {
 	return (
 		reversing ?
 			extract_trailer_index(cstr + 1, *(cstr + 1) != '\0', count + 1) :
